Move supporting functions to functions.py
